Bugfix no BufferedRouter. Estava trasando mensagem interna como se fosse pacote.

+++ src/BufferedRouter.h
@@ -30,7 +30,7 @@ class BufferedRouter : public cSimpleModule
 		cMessage *next;
 		bool isFinite;
 		long bufferSize;
-		int bufferedSize;
+		long bufferedSize;
 
 	protected:
 		virtual void initialize();
